fix: changing implementation classes from public to internal
I'm going to try and use the dependency inversion principle to only expose the public interfaces for implementations that don't require to interact directly with any source projects, like a ASP.NET API project as an example. The only class implementations that will be public besides DI containers will be controller like classes.